Implement new step 3 in create account user flow to add resume, refactoring other components & progress bar logic appropriately
