+++ frontend/src/App.css
@@ -289,7 +289,7 @@ footer {
   list-style-type: none;
   padding: 0;
   margin: 0 0 10px 0;
-  width: 75vw;
+  width: 90vw;
 }
 
 .create-account-progress-bar-content-item {
@@ -302,7 +302,7 @@ footer {
 }
 
 .create-account-progress-bar {
-  width: 75vw;
+  width: 90vw;
   height: 30%;
   border: 2px solid var(--gray);
   border-radius: 20px;
